Refactor database CLI
